# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A tunnel service for exposing local services to the internet",
        "title": "ShipIt Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ShipIt Support",
            "url": "http://www.shipit.com/support",
            "email": "support@shipit.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tunnels": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of tunnels belonging to the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "List user tunnels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by tunnel status (active, inactive, terminated)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of tunnels to return (default: 50, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of tunnels to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of tunnels",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new HTTP or TCP tunnel for exposing local services",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "Create a new tunnel",
                "parameters": [
                    {
                        "description": "Tunnel creation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTunnelRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tunnel created successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateTunnelResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or tunnel limit reached",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Subdomain already taken",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tunnels/{tunnel_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific tunnel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "Get tunnel details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tunnel ID",
                        "name": "tunnel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tunnel details",
                        "schema": {
                            "$ref": "#/definitions/handlers.TunnelListResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Tunnel not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific tunnel. This will terminate the tunnel and close all connections.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "Delete tunnel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tunnel ID",
                        "name": "tunnel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tunnel deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Tunnel not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tunnels/{tunnel_id}/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed statistics for a specific tunnel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tunnels"
                ],
                "summary": "Get tunnel statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tunnel ID",
                        "name": "tunnel_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "24h",
                        "description": "Time period for statistics",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "requests,bandwidth,latency",
                        "description": "Metrics to include",
                        "name": "metrics",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tunnel statistics",
                        "schema": {
                            "$ref": "#/definitions/handlers.TunnelStatsResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Tunnel not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateTunnelRequest": {
            "type": "object",
            "required": [
                "local_port",
                "protocol"
            ],
            "properties": {
                "local_port": {
                    "description": "Port on client side",
                    "type": "integer",
                    "minimum": 1
                },
                "protocol": {
                    "description": "\"http\" or \"tcp\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handlers.TunnelProtocol"
                        }
                    ]
                },
                "subdomain": {
                    "description": "Optional custom subdomain",
                    "type": "string"
                }
            }
        },
        "handlers.CreateTunnelResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "ISO timestamp",
                    "type": "string"
                },
                "protocol": {
                    "description": "http or tcp",
                    "type": "string"
                },
                "public_port": {
                    "description": "For TCP tunnels only",
                    "type": "integer"
                },
                "public_url": {
                    "description": "Public URL for accessing tunnel",
                    "type": "string"
                },
                "status": {
                    "description": "Current status",
                    "type": "string"
                },
                "tunnel_id": {
                    "description": "UUID of created tunnel",
                    "type": "string"
                }
            }
        },
...
        "handlers.TunnelListResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "local_port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                },
                "public_port": {
                    "type": "integer"
                },
                "public_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subdomain": {
                    "type": "string"
                },
                "tunnel_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TunnelStatsResponse": {
            "type": "object",
            "properties": {
                "active_connections": {
                    "type": "integer"
                },
                "analytics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.TunnelAnalyticsPoint"
                    }
                },
                "total_bytes_in": {
                    "type": "integer"
                },
                "total_bytes_out": {
                    "type": "integer"
                },
                "total_requests": {
                    "type": "integer"
                },
                "tunnel_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key authentication for CLI agents and web users",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token authentication for web users",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}